refactor(clipboard): move initCopyContent back to features/copycontent.ts
It's specialized for the repo file view, not a generic clipboard helper.
